requirements.txt
argh==0.26.2
asn1crypto==0.22.0
astroid==1.5.3
backports.functools-lru-cache==1.4
bcrypt==3.1.3
boto3==1.4.4
botocore==1.5.75
cffi==1.10.0
configparser==3.5.0
docutils==0.13.1
enum34==1.1.6
futures==3.1.1
gitdb2==2.0.2
GitPython==2.1.5
idna==2.5
ipaddress==1.0.18
isort==4.2.15
jmespath==0.9.3
lazy-object-proxy==1.3.1
mccabe==0.6.1
meld3==1.0.2
pathtools==0.1.2
pyasn1==0.2.3
pycparser==2.18
pylint==1.7.2
python-dateutil==2.6.0
PyYAML==3.12
s3transfer==0.1.10
singledispatch==3.4.0.3
six==1.10.0
smmap2==2.0.3
supervisor==3.3.2
watchdog==0.8.3
wrapt==1.10.10
